feat(jdbc): add a where clause when fetching max offset
Also some refactoring to reduce code duplication
