+++ jdbc-h2/src/main/java/io/kestra/runner/h2/H2Queue.java
@@ -16,36 +16,8 @@ public class H2Queue<T> extends JdbcQueue<T> {
     }
 
     @Override
-    protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, Integer offset, boolean forUpdate) {
-        var select = ctx.select(
-                AbstractJdbcRepository.field("value"),
-                AbstractJdbcRepository.field("offset")
-            )
-            .from(this.table)
-            .where(AbstractJdbcRepository.field("type").eq(this.cls.getName()));
-
-        if (consumerGroup != null) {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
-        } else {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").isNull());
-        }
-
-        if (offset != 0) {
-            select = select.and(AbstractJdbcRepository.field("offset").gt(offset));
-        }
-
-        var limitSelect = select
-            .orderBy(AbstractJdbcRepository.field("offset").asc())
-            .limit(configuration.getPollSize());
-        ResultQuery<Record2<Object, Object>> configuredSelect = limitSelect;
-
-        if (forUpdate) {
-            configuredSelect = limitSelect.forUpdate().skipLocked();
-        }
-
-        return configuredSelect
-            .fetchMany()
-            .getFirst();
+    protected Condition buildTypeCondition(String type) {
+        return AbstractJdbcRepository.field("type").eq(type);
     }
 
     @Override
