+++ jdbc-mysql/src/main/java/io/kestra/runner/mysql/MysqlQueue.java
@@ -23,36 +23,8 @@ public class MysqlQueue<T> extends JdbcQueue<T> {
     }
 
     @Override
-    protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, Integer offset, boolean forUpdate) {
-        var select = ctx.select(
-                AbstractJdbcRepository.field("value"),
-                AbstractJdbcRepository.field("offset")
-            )
-            .from(this.table)
-            .where(AbstractJdbcRepository.field("type").eq(this.cls.getName()));
-
-        if (offset != 0) {
-            select = select.and(AbstractJdbcRepository.field("offset").gt(offset));
-        }
-
-        if (consumerGroup != null) {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
-        } else {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").isNull());
-        }
-
-        var limitSelect = select
-            .orderBy(AbstractJdbcRepository.field("offset").asc())
-            .limit(configuration.getPollSize());
-        ResultQuery<Record2<Object, Object>> configuredSelect = limitSelect;
-
-        if (forUpdate) {
-            configuredSelect = limitSelect.forUpdate().skipLocked();
-        }
-
-        return configuredSelect
-            .fetchMany()
-            .getFirst();
+    protected Condition buildTypeCondition(String type) {
+        return AbstractJdbcRepository.field("type").eq(type);
     }
 
     @Override
