+++ jdbc/src/main/java/io/kestra/jdbc/runner/JdbcQueue.java
@@ -22,12 +22,8 @@ import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.transaction.exceptions.CannotCreateTransactionException;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.jooq.DSLContext;
-import org.jooq.Field;
-import org.jooq.JSONB;
+import org.jooq.*;
 import org.jooq.Record;
-import org.jooq.Result;
-import org.jooq.Table;
 import org.jooq.impl.DSL;
 
 import java.io.IOException;
@@ -173,7 +169,37 @@ public abstract class JdbcQueue<T> implements QueueInterface<T> {
         return this.receiveFetch(ctx, consumerGroup, offset, true);
     }
 
-    abstract protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, Integer offset, boolean forUpdate);
+    protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, Integer offset, boolean forUpdate) {
+        var select = ctx.select(
+                AbstractJdbcRepository.field("value"),
+                AbstractJdbcRepository.field("offset")
+            )
+            .from(this.table)
+            .where(buildTypeCondition(this.cls.getName()));
+
+        if (offset != 0) {
+            select = select.and(AbstractJdbcRepository.field("offset").gt(offset));
+        }
+
+        if (consumerGroup != null) {
+            select = select.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
+        } else {
+            select = select.and(AbstractJdbcRepository.field("consumer_group").isNull());
+        }
+
+        var limitSelect = select
+            .orderBy(AbstractJdbcRepository.field("offset").asc())
+            .limit(configuration.getPollSize());
+        ResultQuery<Record2<Object, Object>> configuredSelect = limitSelect;
+
+        if (forUpdate) {
+            configuredSelect = limitSelect.forUpdate().skipLocked();
+        }
+
+        return configuredSelect
+            .fetchMany()
+            .getFirst();
+    }
 
     protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, String queueType) {
         return this.receiveFetch(ctx, consumerGroup, queueType, true);
@@ -183,18 +209,26 @@ public abstract class JdbcQueue<T> implements QueueInterface<T> {
 
     abstract protected void updateGroupOffsets(DSLContext ctx, String consumerGroup, String queueType, List<Integer> offsets);
 
+    protected abstract Condition buildTypeCondition(String type);
+
     @Override
     public Runnable receive(String consumerGroup, Consumer<Either<T, DeserializationException>> consumer, boolean forUpdate) {
         AtomicInteger maxOffset = new AtomicInteger();
 
         // fetch max offset
         dslContextWrapper.transaction(configuration -> {
-            Integer integer = DSL
+            var select = DSL
                 .using(configuration)
                 .select(DSL.max(AbstractJdbcRepository.field("offset")).as("max"))
                 .from(table)
-                .fetchAny("max", Integer.class);
+                .where(buildTypeCondition(this.cls.getName()));
+            if (consumerGroup != null) {
+                select = select.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
+            } else {
+                select = select.and(AbstractJdbcRepository.field("consumer_group").isNull());
+            }
 
+            Integer integer = select.fetchAny("max", Integer.class);
             if (integer != null) {
                 maxOffset.set(integer);
             }
